feat: Add `run_for_branch` option (#8)
We want to be able to run the action on release branches as well.
This adds a new option` run_for_branch` which can be set to overwrite the default `isMainBranch` detection.

When this is set to `'true'`, it will run size limit for the given branch without any comparison. Set it to unset or `''` to revert to the default main branch detection.